pare ca am fixat security si dialog, era de la widget wrong word
